AI/requirements.txt
# Core FastAPI and web dependencies
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.0

# LangChain and AI dependencies
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-text-splitters==0.3.8
langchain-chroma==0.2.4

# Vector database
chromadb>=1.0.9

# Document processing
PyPDF2==3.0.1
beautifulsoup4==4.13.4
youtube-transcript-api

# Embeddings and ML
sentence-transformers==5.0.0
transformers==4.53.1
torch>=2.0.0
numpy>=1.26.4

# HTTP and API
requests==2.28.2
httpx==0.28.1

# Wikipedia API
wikipedia==1.4.0

# Utilities
tqdm==4.65.2
typing-extensions==4.13.2

rapidfuzz

